serialize wait and destroy in containerizer
Containerizers keep track of all the containers they have created.  When
a container is destroy()'ed the containerizer removes the container from
its internal hashmap.  This means that wait() must be called before the
container is destroyed.

When the docker containerizer fails because of a "docker pull" or
"docker run" failure, we end up with a race condition where wait() does
not properly return the Termination, so we ended up with a default
failure message of "Abnormal Executor Termination" which isn't very
useful.

This commit solves this issue by not destroying the container until
after wait is called in the failure case.

Fixes MESOS-2656

Review: https://reviews.apache.org/r/33249/